+++ packages/clients/meldemichel/src/mapConfigurations.ts
@@ -102,7 +102,6 @@ const addressSearch: AddressSearchConfiguration = {
 }
 
 const commonPins: Partial<PinsConfiguration> = {
-  // TODO must work separate to GFI (clicking !== selecting, in Meldemichel)
   toZoomLevel: 7,
   movable: true,
   style: {
@@ -131,6 +130,9 @@ const mapConfigurations: Record<
   [MODE.COMPLETE]: (reportServiceId: string, afmUrl: string) => {
     return {
       ...commonMapConfiguration,
+      extendedMasterportalapiMarkers: {
+        layers: [reportServiceId],
+      },
       addressSearch,
       layers: [
         ...commonLayers,
@@ -160,15 +162,24 @@ const mapConfigurations: Record<
       geoLocation,
       gfi: {
         mode: 'bboxDot',
+        activeLayerPath: 'plugin/layerChooser/activeMaskIds',
         layers: {
           [reportServiceId]: {
-            // TODO doesn't work atm; no coordinate source
             geometry: false,
             window: true,
-            properties: { filename: 'Name of file' },
+            // translation in meldemichel's local gfi override
+            properties: [
+              'str',
+              'hsnr',
+              'pic',
+              'skat',
+              'beschr',
+              'rueck',
+              'start',
+              'statu',
+            ],
           },
         },
-        coordinateSources: [], // to be set in addPlugins.ts
       },
       pins: commonPins,
       reverseGeocoder,