+++ packages/clients/meldemichel/src/plugins/Gfi/index.ts
@@ -1,0 +1,3 @@
+import MeldemichelGfiFeature from './Feature.vue'
+
+export default MeldemichelGfiFeature