+++ packages/plugins/Gfi/src/components/FeatureSwitchButtons.vue
@@ -30,7 +30,7 @@ import { mapGetters, mapMutations } from 'vuex'
 type GfiIndexStep = -1 | 1
 
 export default Vue.extend({
-  name: 'GfiSwitchButtons',
+  name: 'GfiFeatureSwitchButtons',
   computed: {
     ...mapGetters('plugin/gfi', [
       'windowFeatures',